move user-group service methods to UserGroupDetail dir
