+++ apps/app/src/components/Admin/UserGroupDetail/use-user-group-resource.ts
